* UPD: Removed network code.

+++ code/engine.vc2008/xrEngine/Environment_misc.cpp
@@ -457,7 +457,7 @@ void	CEnvironment::mods_load()
 		u32 ver = 0x0015;
 		u32 sz;
 
-		while (0 != (sz = fs->find_chunk(id)))
+		while (0 != (sz = u32(fs->find_chunk(id))))
 		{
 			if (id == 0 && sz == sizeof(u32))
 			{
